fix: serialization error directDependencies
The "directDependency" String from the response of project/lookup can be
null.
to fix the error, we make this attribute nullable and warn the user in
GetOutdatedDependenciesTask if the project has no direct dependencies.

also fix an issue in the riskScore task, when the sbom analyse failed
and the project then has no direct dependencies.